gmock: fixed the build by introducing a workaround around an issue in build2
As for the buildfile in gtest, there is a workaround
for targets from libx{whatever} which are not defined
even if we defined lib{whatever}. The workaround
is to define them without requirements.

This is an issue I spotted in the past, but couldnt
report without more information and without
trying to remake these packages using links.
(see notes for history).